move calibrate logic
